Incrase size of balance field to handle current balance

+++ mtp_api/apps/account/models.py
@@ -5,7 +5,7 @@ from transaction.utils import format_amount
 
 
 class Balance(TimeStampedModel):
-    closing_balance = models.IntegerField()
+    closing_balance = models.BigIntegerField()
     date = models.DateField()
 
     class Meta:
